refactor: Core 인터페이스 폴더(네임스페이스) 이름을 알아보기 쉽게 변경
