+++ src/MSD.Crux/MSD.Crux.API/Controllers/UserAndEmployee/EmployeeController.cs
@@ -1,8 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 using MSD.Crux.API.Helpers;
+using MSD.Crux.Core.IServices;
 using MSD.Crux.Core.Models;
-using MSD.Crux.Core.Services;
-using MSD.Crux.Core.Services;
 
 namespace MSD.Crux.API.Controllers;
