+++ src/MSD.Crux/MSD.Crux.Core/IServices/IUserLoginService.cs
@@ -1,6 +1,6 @@
 using MSD.Crux.Core.Models;
 
-namespace MSD.Crux.Core.Services;
+namespace MSD.Crux.Core.IServices;
 
 /// <summary>
 /// 유저 로그인 ID/PW 생성과  로그인시 JWT 발급 관련 서비스 인터페이스
