+++ src/MSD.Crux/MSD.Crux.Host/Program.cs
@@ -7,8 +7,8 @@ using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.IdentityModel.Tokens;
 using MSD.Crux.API.Helpers;
 using MSD.Crux.Core.Helpers;
-using MSD.Crux.Core.Repositories;
-using MSD.Crux.Core.Services;
+using MSD.Crux.Core.IRepositories;
+using MSD.Crux.Core.IServices;
 using MSD.Crux.Infra.Repositories;
 using MSD.Crux.Infra.Services;
 using Npgsql;
