--- src/MSD.Crux/MSD.Crux.Infra/Class1.cs
@@ -1,5 +1,0 @@
-﻿namespace MSD.Crux.Infra;
-
-public class Class1
-{
-}
