+++ src/MSD.Crux/MSD.Crux.Infra/Repositories/Db/EmployeeRepoPsqlDb.cs
@@ -1,8 +1,8 @@
 using System.Data;
 using Dapper;
 using MSD.Crux.Core.Helpers;
+using MSD.Crux.Core.IRepositories;
 using MSD.Crux.Core.Models;
-using MSD.Crux.Core.Repositories;
 
 namespace MSD.Crux.Infra.Repositories;
