+++ src/MSD.Crux/MSD.Crux.Infra/Repositories/InMemory/EmployeeRepoInMemory.cs
@@ -1,7 +1,7 @@
 using System.Collections.Concurrent;
 using MSD.Crux.Core.Helpers;
+using MSD.Crux.Core.IRepositories;
 using MSD.Crux.Core.Models;
-using MSD.Crux.Core.Repositories;
 
 namespace MSD.Crux.Infra.Repositories;
