+++ src/MSD.Crux/MSD.Crux.Infra/Repositories/InMemory/UserRepoInMemory.cs
@@ -1,6 +1,6 @@
 using System.Collections.Concurrent;
+using MSD.Crux.Core.IRepositories;
 using MSD.Crux.Core.Models;
-using MSD.Crux.Core.Repositories;
 
 namespace MSD.Crux.Infra.Repositories;
