+++ src/MSD.Crux/MSD.Crux.Infra/Services/Default/UserLoginService.cs
@@ -1,8 +1,8 @@
 using Microsoft.Extensions.Configuration;
 using MSD.Crux.Core.Helpers;
+using MSD.Crux.Core.IRepositories;
+using MSD.Crux.Core.IServices;
 using MSD.Crux.Core.Models;
-using MSD.Crux.Core.Repositories;
-using MSD.Crux.Core.Services;
 
 namespace MSD.Crux.Infra.Services;
